Refactor file paths in BIDS project configuration and update common BIDS config template
